requirements.txt
Django==4.2.*
environs==9.5.*
django-multiselectfield==0.1.*
django-admin-rangefilter==0.10.*
